Moved BinaryArray to its own module.
