+++ rust/arrow/src/array/equal_json.rs
@@ -17,9 +17,7 @@
 
 use super::*;
 use crate::datatypes::*;
-use array::{
-    Array, BinaryOffsetSizeTrait, GenericBinaryArray, GenericListArray, OffsetSizeTrait,
-};
+use array::{Array, GenericListArray, OffsetSizeTrait};
 use hex::FromHex;
 use serde_json::value::Value::{Null as JNull, Object, String as JString};
 use serde_json::Value;
