+++ rust/arrow/src/array/mod.rs
@@ -83,6 +83,7 @@
 
 #[allow(clippy::module_inception)]
 mod array;
+mod array_binary;
 mod array_dictionary;
 mod array_string;
 mod array_struct;
@@ -107,13 +108,13 @@ pub use self::data::ArrayData;
 pub use self::data::ArrayDataBuilder;
 pub use self::data::ArrayDataRef;
 
-pub use self::array::BinaryArray;
-pub use self::array::FixedSizeBinaryArray;
 pub use self::array::FixedSizeListArray;
-pub use self::array::LargeBinaryArray;
 pub use self::array::LargeListArray;
 pub use self::array::ListArray;
 pub use self::array::PrimitiveArray;
+pub use self::array_binary::BinaryArray;
+pub use self::array_binary::FixedSizeBinaryArray;
+pub use self::array_binary::LargeBinaryArray;
 pub use self::array_dictionary::DictionaryArray;
 pub use self::array_string::LargeStringArray;
 pub use self::array_string::StringArray;
@@ -161,9 +162,10 @@ pub type DurationMillisecondArray = PrimitiveArray<DurationMillisecondType>;
 pub type DurationMicrosecondArray = PrimitiveArray<DurationMicrosecondType>;
 pub type DurationNanosecondArray = PrimitiveArray<DurationNanosecondType>;
 
-pub use self::array::GenericBinaryArray;
 pub use self::array::GenericListArray;
 pub use self::array::OffsetSizeTrait;
+pub use self::array_binary::BinaryOffsetSizeTrait;
+pub use self::array_binary::GenericBinaryArray;
 pub use self::array_string::GenericStringArray;
 pub use self::array_string::StringOffsetSizeTrait;
